interferogram and spectra module
